+++ config/medium/TicTacToeEngine/code/CGCode.cpp
@@ -1,0 +1,21 @@
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+int main()
+{
+    string engine;
+    getline(cin, engine);
+    for (int i = 0; i < 3; i++) {
+        string row;
+        getline(cin, row);
+    }
+
+    // Write an answer using cout. DON'T FORGET THE "<< endl"
+    // To debug: cerr << "Debug messages..." << endl;
+
+    cout << "XOXO" << endl;
+}