+++ config/medium/TicTacToeEngine/code/CGCode.ts
@@ -1,0 +1,9 @@
+const engine: string = readline();
+for (let i = 0; i < 3; i++) {
+    const row: string = readline();
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('XOXO');