+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - configuration for "Lines Intersections".
+- configuration for "Tic Tac Toe Engine".
 
 ## [1.21.0] - 2025-11-30
 ### Added